Refactor how hw components reference each other
Instead of injecting the various links into each constructor (and sometimes
via an additional connect() method when components rely on each other,
the components just recieve a reference to the SpelBoy object and can obtain
references from there. Getters are then used to maintain the existing interface.

This couples the components more strongly together, but I think the tradeoff will
be worth it